Remove the lowperformance build tag
We don't have anymore performance issues with the scalingstack instances,
but were using this tag to prevent the execution of the Ubuntu Fan suite,
which was flaky in low performance machines and now was failing because
of a restrictive firewall that prevented the suite from connecting to
Docker Hub. Now we have an additional problem,
https://bugs.launchpad.net/snappy/+bug/1544507, and the tests have been
skipped, so no need to keep the build tag.